server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
fastapi==0.135.3
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0

# Validations
pydantic[email]

# ORM and database connection
sqlalchemy
alembic
psycopg2-binary

# Environment Variables
python-dotenv>=1.0.0

# Password hashing
bcrypt

# JWT
python-jose[cryptography]

# Tests
pytest
pytest-asyncio

# Code quality
ruff
mypy